- Added IdGenerator. Automatic id generation through sequence shall be gradually removed. This is to support return the id of inserted entity, which eventually will be used to index in solr. - Initially refactored CourseTeacher, Examiner and AdditionalRolePermission to use IdGenerator.

+++ ums-api/src/main/java/org/ums/domain/model/mutable/MutableAssignedTeacher.java
@@ -8,7 +8,7 @@ import org.ums.domain.model.immutable.Semester;
 import org.ums.domain.model.immutable.Teacher;
 
 public interface MutableAssignedTeacher extends AssignedTeacher, Mutable, MutableLastModifier,
-    MutableIdentifier<Integer> {
+    MutableIdentifier<Long> {
 
   void setSemester(final Semester pSemester);
